requirements (1).txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Redis & Task Queue
redis==5.0.1
celery==5.3.4

# Music APIs
spotipy==2.23.0
lyricsgenius==3.0.1
requests==2.31.0

# NLP & ML
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0
sentence-transformers==2.2.2
vaderSentiment==3.3.2
spacy==3.7.2
nltk==3.8.1

# Language Detection
fasttext==0.9.2
langdetect==1.0.9

# Audio Processing (optional)
librosa==0.10.1
audioread==3.0.1

# Time Series
statsmodels==0.14.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Visualization & Reporting
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
